More updates for rollup

+++ charts/tezos/scripts/smart-rollup-snapshot-importer.sh
@@ -4,14 +4,20 @@ bin_dir="/usr/local/bin"
 data_dir="/var/tezos"
 smart_rollup_node_dir="${data_dir}/rollup"
 smart_rollup_node="${bin_dir}/octez-smart-rollup-node"
-snapshot_file="${rollup_rollup_dir}/rollup.snapshot"
+smart_rollup_snapshot_file="${smart_rollup_node_dir}/rollup-snapshot.archive"
 
 if [ ! -f ${snapshot_file} ]; then
     echo "No snapshot to import."
     exit 0
 fi
 
-${smart_rollup_node} snapshot import ${snapshot_file} --data-dir ${smart_rollup_node_dir} --no-check
-find ${node_dir}
+if [ -e "${smart_rollup_node_dir}/context/store.dict" ]; then
+  echo "Smart rollup snapshot has already been imported. Exiting."
+  exit 0
+fi
+
+echo "Importing smart rollup snapshot"
+${smart_rollup_node} snapshot import ${smart_rollup_snapshot_file} --data-dir ${smart_rollup_node_dir} --no-check
+find ${smart_rollup_node_dir}
 
 rm -rvf ${snapshot_file}
